rename parsedInputSequence.ts to blastSequence.ts
