+++ src/content/app/tools/blast/types/blastSequence.ts
